+++ src/main/java/com/refactoringhabit/auth/controller/AuthRestController.java
@@ -9,10 +9,12 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
@@ -22,6 +24,11 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthRestController {
 
     private final AuthService authService;
+
+    @GetMapping("/check-email")
+    public ApiResponse<Boolean> checkEmail(@RequestParam("email") String email) {
+        return ApiResponse.ok(authService.emailCheck(email));
+    }
 
     @PostMapping("/find-email")
     public ApiResponse<String> findEmail(@RequestBody FindEmailRequestDto findEmailRequestDto) {
