+++ src/main/java/com/refactoringhabit/member/controller/MemberRestController.java
@@ -5,10 +5,8 @@ import com.refactoringhabit.member.domain.service.MemberService;
 import com.refactoringhabit.member.dto.MemberJoinRequestDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
@@ -28,10 +26,5 @@ public class MemberRestController {
     ) {
         memberService.memberJoin(memberJoinRequestDto, multipartFile);
         return ApiResponse.created();
-    }
-
-    @GetMapping("/check-email")
-    public ApiResponse<Boolean> checkEmail(@RequestParam("email") String email) {
-        return ApiResponse.ok(memberService.emailCheck(email));
     }
 }
